Break up main server components
