+++ dispute_management_service/src/main/scala/routes/package.scala
@@ -1,0 +1,14 @@
+import cats.effect.IO
+import org.http4s.{HttpApp, HttpRoutes}
+import org.http4s.dsl.io._
+import org.http4s.implicits._
+
+package object routes {
+
+  def httpApp: HttpApp[IO] = {
+    HttpRoutes.of[IO] {
+      case GET -> Root / "ping" =>
+        Ok("PONG")
+    }.orNotFound
+  }
+}
